🚀 [Feat]: Apply suspense & error-boundary to bookmark
북마크 기능에 대해 suspense와 Error-boundary를 적용함.
기존에 page에서 template으로 필요한 데이터를 내려주는 것이 필요한 component
내부로 옮김(suspense 적용을 위함)
